Extract functions and classes from __init__.py into an own python file
